# src/main/resources/api.yaml
openapi: 3.0.3
info:
  title: Bookstore Api
  description: Bookstore Api.
  version: 0.0.1
paths:
...
  /carts/{user_id}/items:
    get:
      description: Get list of items from cart.
      operationId: getCartContentById
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: integer
            example: 1
      responses:
        200:
          description: Users cart contents.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Book"
        401:
          description: User not logged in.
        403:
          description: User not authorized.
    delete:
      description: Clear the cart.
      operationId: clearCartContentById
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: integer
            example: 1
      responses:
        204:
          description: Cart cleared.
        400:
          description: Bad request.
        401:
          description: User not logged in.
        403:
          description: User not authorized.
  /carts/{user_id}/items/{item_id}:
    put:
      description: Add an items to cart.
      operationId: addItemToCartById
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: integer
            example: 1
        - name: item_id
          in: path
          required: true
          schema:
            type: integer
            example: 1
      responses:
        204:
          description: Item added to cart.
        400:
          description: Bad request.
        401:
          description: User not logged in.
        403:
          description: User not authorized.
        404:
          description: Item not found.
    delete:
      description: Remove item from cart.
      operationId: removeItemFromCartById
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: integer
            example: 1
        - name: item_id
          in: path
          required: true
          schema:
            type: integer
            example: 1
      responses:
        204:
          description: Item removed from cart.
        400:
          description: Bad request.
        401:
          description: User not logged in.
        403:
          description: User not authorized.
        404:
          description: Item not present in cart.
  # Order, order states
...
components:
  schemas:
    Book:
      type: object
      properties:
        id:
          type: integer
          example: 1
          readOnly: true
        title:
          type: string
          example: "Guts 'N Gunships"
        author:
          type: string
          example: "Mark Garrison"
        publisher:
          type: string
          example: "Mark Garrison"
...